Update MedicalDataSource's FhirVersion to be NonNull
After some discussion, we still decided that this should be NonNull. The
reasons are as below:

1)
In the short-term MVP launch, we would like to force that one data source
MUST be limited to accept only ONE FHIR version.

Thus, we have created a REQUIRED field in:
CreateMedicalDataSourceRequest.Builder(fhirBaseUri, displayName, fhirVersion)

It is then very confusing that FhirVersion in MedicalDataSource is nullable.

With this fix, MedicalDataSource.getFhirVersion() returns the above fhirVersion
set by the users. This is consistent.

2)
In the long-term future, we may or may not remove the above limitation.

If we do, there're already discussions in ag/29358463, that the plan is
to create another constructor without fhirVersion:
CreateMedicalDataSourceRequest.Builder(fhirBaseUri, displayName)

We could also create a constant representing "no limitation": NON_RESTRICTED_FHIR_VERSION.
So that when the new construcor is called, fhirVersion will be set to
this NON_RESTRICTED_FHIR_VERSION.

we cannot forbid users to call builder.setFhirVersion after using the
new builder, but this is fine.

And with this fix, MedicalDataSource.getFhirVersion() can also return
this NON_RESTRICTED_FHIR_VERSION. This is also consistent.

Conclusion, this fix works both for short-term and long-term. And given
that we don't know yet if 2) long-term future would ever happen, it is
important that we try our best to make 1) short-term API more consistent.

Bug: 365958816
Test: atest
Flag: com.android.healthfitness.flags.personal_health_record
Change-Id: I67bd0987238c60bba5c44f77b88ffc446cae2058