Fix a bug leading to deadlock in CTS tests.
AppInfoHelper is synchronized, but obtains a SQLite connection from the connection pool inside a synchronized block.
This means if one of its synchronized method is called from within a TransactionManager SQLite transaction it can lead to deadlock.

This CL is not a long term fix, but mitigates the short term problem while a better plan is considered.

Bug: http://b/358274374
Flag: com.android.healthfitness.flags.personal_health_record
Test: atest -c --test-timeout 20000 CtsHealthFitnessDeviceTestCases:HealthConnectManagerTest (multiple devices, multiple runs)
Change-Id: I4fed0109e781b75eea8e740531c9d68bcab48f9b