Refactor code so that UsageStatsLogger does not rely on HealthConnectPermissionHelper.
UsageStatsLogger needed the HealthConnectPermissionHelper to get
permission granted to each connected app. The function in
HealthConnectPermissionHelper being used had no dependencies however
cannot be made static so we had initialize HealthConnectPermissionHelper
with a bunch of dependencies causing unnecessary complications. To solve
this, I copied the function being used in HealthConnectPermissionHelper
to the UsageStatsCollector.

Test: atest
Flag: EXEMPT refactoring
Bug: 368069094
Change-Id: I4c488bc58be3e25f65fadb6d5b0f6f6e1a685700