Reject contained resources during upsert of FHIR resources
Even though the contained field is valid FHIR we don't want to allow this in the first version as it brings some
complication related to categorisation and permissioning which need more
discussion.

This change allows the contained field to be set to an empty array for
now, but the FhirResourceValidator may disallow empty arrays later on when
array validation is implemented.

Bug: 376463270
Test: atest
Flag: com.android.healthfitness.flags.phr_fhir_structural_validation

Change-Id: I4eb55d4512ea96d7f31491b63c7967be114743a3