HealthConnect: Require permission to get HealthConnectManager on Wear.
We need HealthConnectManager available for internal health connect use for permissions requesting and managing but don't want to expose the full set of APIs to 3P apps. In order to do this we will only return the full HealthConnectManager to apps that have the MANAGE_HEALTH_PERMISSIONS permission, otherwise we continue to return null.

Bug: 372243727
Flag: android.permission.flags.replace_body_sensor_permission_enabled
Test: atest HealthServicesInitializerTest passes on Wear emulator with flags enabled. Local test passes with WHS TestApp being granted health permissions but not being able to get a reference to HealthConnectManager: https://paste.googleplex.com/5417065963061248
Change-Id: Icf15ea09a8f67541af5602cc39876fc2c08256f3