Introduce context+user as a constructor param in AppInfoHelper.
In the past, AppInfoHelper was attempted to be written to support multiple users at the same time.
But the underlying storage (in TransactionManager) and the internal cache only
have a single instance, and are not capable of working on multiple users
at the same time.

Some callers were passing in the right user context (like
DataMigrationHelper), but others were just always passing in the wrong
user context (like UpsertTransactionRequest).  Where the wrong context
was passed in, things didn't work as expected.

Test: presubmits
Flag: NONE no-op for primary user(id 0). For secondary users, see comments.
Bug: 375168249
Bug: 371122910
Bug: 375351821
Change-Id: I79926578bef96f0d3e887e121eef5f651f8071c8