Validate structure of nested types
This is a second version of ag/31176212, without using recursion to
achieve the validation.

FHIR doesn't set a maximum level of nesting that is allowed, but it
seems like in practice we shouldn't expect data that is nested too
deeply. Therefore we are setting a max nesting level to 20, to avoid
consuming too much memory if an object is very deeply nested.

Test: atest
Bug: 361775175
Fixes: 361775175
Flag: com.android.healthfitness.flags.phr_fhir_complex_type_validation
Change-Id: I8acc225f590741c1ca9aea2aeeda687408102f16