Use IntroPreference instead of AppHeaderPreference when expressive theming is on
IntroPreference: https://hsv.googleplex.com/6283437953515520

Ignored some tests for now that I'll fix separately. Basically
IntroPreference is a bit higher than AppHeaderPreference so Additional
access button does not fit to the screen and it cannot be found for some
reason despite the scrollTo. Some tests are already ignored because of
this, so not this change introduces the issue.

Fix: 377663414
Test: manual
Flag: NONE covered by settings
Change-Id: I005943f68a979c049a69293a00c34174a2121cd4