Fix series data aggregations for Steps Cadence
Currently, the interval associated with the entire SeriesRecord is used
for aggregation. This changes this to use the individual series sample
timestamps.

This cl does not fix this issue for local time filters, where there is
additional complexity due to the fact that we do not store sample
timestamps in local time in the same way that we do for the interval
timestamps. This will require a bigger refactor, which should be enabled
by future work to fix b/379052468.

This needs to be changed for all series data types, but further data
types will be added in follow up cls

Bug: 326058390
Test: AggregationApisTest
Flag: EXEMPT bugfix

BYPASS_NULLAWAY_CHECK_REASON=match existing pattern

Change-Id: Iaccbd8bbbaf805afb3b70207292e382b36b26c1d