Turn error codes and state indicators into one state machine for import progress
This simplifies the code and makes it less error prone, because the state of an import is tracked in one consistent way and an import is always in one well defined state (started, error or success). Replaces the previous implementation where we were trackign separately if an import is ongoing, which allowed inconsitant states such as ongoing and error at the same time.

Bug: b/356393172
Flag: EXEMPT refactor
Test: atest

Change-Id: Iafe06cdf3c0fa2eaf1c1715c1aca627f8defd015