Introduce RequestContext class for holding state common to all requests.
This encapsulates various things like calling user etc. I've added a few
up front, but will add more in later CLs.

The advantages of this is that it makes it easier to pass these things
down through the call chain. It also handles much of the initialization
in one place, reducing the risk of error.

The unit test has some shortcomings as it doesn't actually run in a
binder thread so isn't very realistic, but at least verifies that
something gets set and the code doesn't blow up for trivial reasons. The
existing CTS tests verify the correct behavior realistically.

Test: unit tests. Impact on existing code covered by existing CTS tests.
Flag: EXEMPT refactor only.
Bug: 381848239
Change-Id: Ia3923a2df707c2565f5d9d5e78e47314184a783b