Fix series data aggregations for PowerRecord
Currently, the interval associated with the entire SeriesRecord is used
for aggregation. This changes this to use the individual series sample
timestamps.

This cl does not fix this issue for local time filters, where there is
additional complexity due to the fact that we do not store sample
timestamps in local time in the same way that we do for the interval
timestamps. This will require a bigger refactor, which should be enabled
by future work to fix b/379052468 .

Bug: 326058390
Test: AggregationApisTest
Flag: EXEMPT bugfix
Change-Id: I9c8953c9a0297bf30bc412476d029dcc76ae5960