Add HealthConnectLibraryAndApplication android_library and filegroup test srcs
This change:
Creates a new android_library to be used within the screenshot tests setup. We need bith the library, but the application class source as well.
Filegroup test srcs to be added in the screenshot tests setup

Both of the additions are used in the change binded with the topic (inside the HealthFitness //vendor/… module)

Test: atest
Fix: 391515484
Flag: NONE Not needed
Change-Id: I4152c1ecb00862fe74691e9db21f2f8fbe92f534