Fix exercise session extra data crash
When exercise sessions are read, there is an extra query that is run to retrieve routes and populate the records with this information. This query, as defined in ExerciseSessionRecordHelper#getExtraDataReadRequests(), does not apply a page size.

This works fine if the routes are filtered by the UIDs of the requested sessions, as that means only the relevant routes are read. If the sessions are read without UIDs, then the query to read the extra data returns all rows that match the time based filtering, without a page size limit, unlike the query for the sessions themselves which does have a page limit.

The method that I updated assumed that a session would exist for every location row that is read, but as explained above there are cases where not all locations that are read have a matching session that was read.

This change checks if the record is in the map before attempting to retrieve it, preventing a crash. Plus a bit of refactoring to make the code easier to understand.

TL;DR: Let's read the sessions with time > -1 with page size 1000, okay 1000 sessions read. Let's also read the routes with time > -1 but no page size, okay 63k locations read. Let's match each of these locations with a session. Oh no there's no session for this location. Boom.

Bug: 394019920
Bug: 396104130
Flag: EXEMPT bugfix
Test: go/testedequals
Change-Id: Icae353035f6d9120297cf3f6483b1aadf5366031