Update the error retry count
Reset it when an export succeeds, when the current error message does
not match the previous error message, when export settings are changed by the user and when
retries finish/a new regular export is scheduled.
Increase it on retries with the same error.
Will be used to remove duplicate errors in our logging.

Test: atest com.android.server.healthconnect.storage
Bug: b/374702524
Flag: com.android.healthfitness.flags.extend_export_import_telemetry
Change-Id: Ia6a0fd71588cef2c85d6816275421ec88f5c2010