WearSettings: Hide background toggle if permission not requested.
If the app doesn't request READ_HEALTH_DATA_IN_BACKGROUND then we
don't want the user to have the illusion of choosing between "all the time" or "while in use"

As part of this change it was discovered that the ViewModel also
checks for supported packages (via the permission rationale intent) in order to load all of the requested permissions. As part of this change
we can opt the Wear apps out of this requirement such that they are
always considered supported. This has the side effect of fixing an AppInfo page bug where non-granted permissions were not being shown.

Change-Id: Ied97e1721ba0ad6b4518d55245beddc7d3bb4fad
Flag: com.android.healthfitness.flags.replace_body_sensor_permission_enabled
Bug: 373692595
Bug: 399218246
Test: Built the emulator and observed the background toggle showing as appropriate. Also saw the ungranted permissions appear in App Info. More work required to get the HealthConnectControllerPermissionTests tests passing on Wear.