AccessLogsHelper: Read logs access from AppOps for granular app ops.
Access to READ_HEART_RATE, READ_SKIN_TEMPERATURE, and READ_OXYGEN_SATURATION all should be tracked via their new AppOps. We need to read out historical access from there as well as suppress the write to the HC access data store (so we don't write redundant data).

Note: We had to add a local wrapper of the AccessLogs class in order to facilitate testing with the "test_current" SDK (needed for simulating the historical app ops). The AccessLogs constructor is hidden and wasn't being found while using test_current.

Note: We are not explicitly using the healthfitness flag for this logic but instead relying on the permission flagging defining the AppOps mapping. If core permissions flag is disabled the AppOps won't be defined and this feature will be turned off.

Bug:364643016
Test: atest HealthFitnessTestSdkUnitTests & atest HealthFitnessUnitTests
Flag: android.permission.flags.replace_body_sensor_permission_enabled
Change-Id: Ief4cc8ad129a9423b29e5d739e9829916db8c3cc