Remove uses of ExtendedMockito for mocking job scheduling by switching to an underlying mock JobScheduler.
This is almost a test-only pure refactor. However, while making the change I noticed the code was
getting all pending jobs and then looping cancelling them individually rather than using cancelAll().
Testing is far easier with cancelAll(), and I believe the code is simpler and more correct, so I made the production code change.
I believe this code is essentially a pure refactor and invisible to any user so I have not flagged it.

This then led to HealthConnectManagerServiceTest failing as it was using the same jobScheduler for all fakes, and so was getting an incorrect cancelAll() count, so I fixed that.

Flag: EXEMPT refactor
Test: atest HealthFitnessUnitTests
Bug: b/396076977
Change-Id: Ia5e8cbfa061f5a486857d65a0d8b708d4d1d5eb5