Get HealthConnectThreadSchedulerTest passing on robolectric.
There are two parts to this. The first is I split one test method that was testing multiple things into multiple methods.
The second is I added some extra mocking code for one of the tests. This is because on device unit tests it is just true that the test is running in the background. But using Robolectric we have to force this with the mock. The code was duplicated from the Foreground case.

Bug: b/403609828
Flag: EXEMPT test only
Test: atest HealthFitnessUnitTests
Change-Id: I2ecfa09cb14a1f6acbfca6fd696221cc239066c8